requirements.txt
huggingface_hub
python-dotenv
requests
chromadb